change double to float

+++ db/migrate/20141117174333_add_ratings_to_tracks.rb
@@ -1,5 +1,5 @@
 class AddRatingsToTracks < ActiveRecord::Migration
   def change
-  	add_column :tracks, :author_rating, :double
+  	add_column :tracks, :author_rating, :float
   end
 end
